아따메 힘들다 그지같은거 어트리뷰트추가 커스텀 경험치관린  추가 명함UI 추가 스탯 추가 랭킹추가

+++ src/main/java/com/ansan/ansanpack/command/DebugStatExpCommand.java
@@ -38,7 +38,7 @@ public class DebugStatExpCommand {
                 Commands.literal("경험치추가")
                         .then(Commands.argument("amount", IntegerArgumentType.integer(1))
                                 .executes(ctx -> {
-                                    int amount = IntegerArgumentType.getInteger(ctx, "amount");
+                                    double amount = IntegerArgumentType.getInteger(ctx, "amount");
                                     LocalPlayerLevelData.INSTANCE.addExp(amount);
 
                                     ctx.getSource().sendSystemMessage(Component.literal("경험치 +" + amount));
