+++ src/main/java/com/ansan/ansanpack/config/LevelDatabaseManager.java
@@ -9,27 +9,27 @@ import java.util.function.BiConsumer;
 
 public class LevelDatabaseManager {
 
-    public static void saveOrUpdate(UUID uuid, String playerName, int level, int exp) {
+    public static void saveOrUpdate(UUID uuid, String playerName, int level, double exp) {
         Properties props = UpgradeConfigManager.loadDbProps();
 
         String url = "jdbc:mysql://" + props.getProperty("db.host") + ":" + props.getProperty("db.port") + "/" +
                 props.getProperty("db.database") + "?serverTimezone=Asia/Seoul&useSSL=false&allowPublicKeyRetrieval=true";
 
         try (Connection conn = DriverManager.getConnection(url, props.getProperty("db.user"), props.getProperty("db.password"))) {
-            // 테이블 없으면 생성
+            // ✅ 테이블 생성 (exp 타입을 DOUBLE로 수정)
             try (Statement tableCheck = conn.createStatement()) {
                 tableCheck.executeUpdate("""
                     CREATE TABLE IF NOT EXISTS player_levels (
                         uuid VARCHAR(36) PRIMARY KEY,
                         player_name VARCHAR(50),
                         level INT NOT NULL,
-                        exp INT NOT NULL,
+                        exp DOUBLE NOT NULL,
                         updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
                     )
                 """);
             }
 
-            // UPSERT 처리
+            // ✅ UPSERT 처리 (exp: setDouble)
             try (PreparedStatement stmt = conn.prepareStatement("""
                 INSERT INTO player_levels (uuid, player_name, level, exp)
                 VALUES (?, ?, ?, ?)
@@ -41,7 +41,7 @@ public class LevelDatabaseManager {
                 stmt.setString(1, uuid.toString());
                 stmt.setString(2, playerName);
                 stmt.setInt(3, level);
-                stmt.setInt(4, exp);
+                stmt.setDouble(4, exp);
                 stmt.executeUpdate();
             }
 
@@ -49,7 +49,8 @@ public class LevelDatabaseManager {
             AnsanPack.LOGGER.error("[LevelDB] 레벨 저장 실패: {}", e.getMessage());
         }
     }
-    public static void load(UUID uuid, BiConsumer<Integer, Integer> callback) {
+
+    public static void load(UUID uuid, BiConsumer<Integer, Double> callback) {
         Properties props = UpgradeConfigManager.loadDbProps();
 
         String url = "jdbc:mysql://" + props.getProperty("db.host") + ":" + props.getProperty("db.port") + "/" +
@@ -62,15 +63,15 @@ public class LevelDatabaseManager {
 
                 if (rs.next()) {
                     int level = rs.getInt("level");
-                    int exp = rs.getInt("exp");
+                    double exp = rs.getDouble("exp");  // ✅ 변경
                     callback.accept(level, exp);
                 } else {
-                    callback.accept(1, 0); // 기본값: 1레벨, 0 경험치
+                    callback.accept(1, 0.0); // 기본값
                 }
             }
         } catch (SQLException e) {
             AnsanPack.LOGGER.error("[LevelDB] 레벨 불러오기 실패: {}", e.getMessage());
-            callback.accept(1, 0); // 실패 시 기본값
+            callback.accept(1, 0.0); // 실패 시 기본값
         }
     }
 }
